fix: build and warnings

+++ libs/libsmlp/src/DataParser.cpp
@@ -102,7 +102,7 @@ Record DataParser::processInputOnly(
     size_t input_size) const {
   std::vector<float> inputs(input_size);
 
-  std::transform(cell_refs.begin(), cell_refs.begin() + input_size,
+  std::transform(cell_refs.begin(), cell_refs.begin() + (int)input_size,
                  inputs.begin(), getFloatValue);
   return {.inputs = inputs, .outputs = std::vector<float>{}};
 }
@@ -113,9 +113,9 @@ Record DataParser::processInputFirst(
   std::vector<float> inputs(input_size);
   std::vector<float> expected_outputs(output_size);
 
-  std::transform(cell_refs.begin(), cell_refs.begin() + input_size,
+  std::transform(cell_refs.begin(), cell_refs.begin() + (int)input_size,
                  inputs.begin(), getFloatValue);
-  std::transform(cell_refs.begin() + input_size, cell_refs.end(),
+  std::transform(cell_refs.begin() + (int)input_size, cell_refs.end(),
                  expected_outputs.begin(), getFloatValue);
   return {.inputs = inputs, .outputs = expected_outputs};
 }
@@ -126,9 +126,9 @@ Record DataParser::processOutputFirst(
   std::vector<float> inputs(input_size);
   std::vector<float> expected_outputs(output_size);
 
-  std::transform(cell_refs.begin(), cell_refs.begin() + output_size,
+  std::transform(cell_refs.begin(), cell_refs.begin() + (int)output_size,
                  expected_outputs.begin(), getFloatValue);
-  std::transform(cell_refs.begin() + output_size, cell_refs.end(),
+  std::transform(cell_refs.begin() + (int)output_size, cell_refs.end(),
                  inputs.begin(), getFloatValue);
   return {.inputs = inputs, .outputs = expected_outputs};
 }
